+++ Tests/test1.csv
@@ -1,0 +1,36 @@
+Questions
+What is LenDenClub?
+How does P2P lending work on LenDenClub?
+Is LenDenClub registered with RBI?
+Is this platform safe to invest in?
+Can I lose my money in P2P lending?
+I signed up but didn’t complete registration. Can you help?
+What documents are needed for registration?
+How do I do CKYC or DigiLocker verification?
+I’m facing an issue during signup. What should I do?
+What happens after I finish KYC?
+How do I start lending on LenDenClub?
+What’s the difference between manual and lumpsum lending?
+How much do I need to start lending?
+Can I choose who to lend to?
+What is the minimum and maximum lending amount?
+When do I start getting returns?
+How are the repayments made to my bank?
+What is the expected return percentage?
+How can I check my earnings?
+Do I get monthly income from lending?
+What if the borrower doesn’t pay back?
+How is risk managed on LenDenClub?
+What is the NPA rate of the platform?
+What is diversification in P2P lending?
+How many borrowers should I lend to?
+How do I add funds to my LenDenClub account?
+Can I use UPI to add funds?
+Is there any fee for depositing or withdrawing money?
+Can I withdraw my money anytime?
+What if I want to pause lending for a while?
+Can I talk to a human agent?
+Where can I report a technical issue?
+I want to delete my account. How do I do that?
+Is there customer support on weekends?
+Can someone help me understand my dashboard?
